requirements.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
distro==1.9.0
filelock==3.29.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
numpy==2.4.4
outcome==1.3.0.post0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-whois==0.9.6
requests==2.33.1
requests-file==3.0.1
scikit-learn==1.8.0
scipy==1.17.1
selenium==4.43.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.1.4
threadpoolctl==3.6.0
tldextract==5.3.1
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.8
wsproto==1.3.2
